requirements.txt
beautifulsoup4
joblib
timidity
transformers
tokenizers
pytorch
accelerate